ton wallet: determine wallet version by cell hash, make cells from wallet boc in init()
