+++ ton/wallet/v1.go
@@ -1,31 +1,6 @@
 package wallet
 
-import "encoding/hex"
-
 // https://github.com/toncenter/tonweb/blob/master/src/contract/wallet/WalletSources.md#v1-wallet
 const _V1R1CodeHex = "B5EE9C72410101010044000084FF0020DDA4F260810200D71820D70B1FED44D0D31FD3FFD15112BAF2A122F901541044F910F2A2F80001D31F3120D74A96D307D402FB00DED1A4C8CB1FCBFFC9ED5441FDF089"
 const _V1R2CodeHex = "B5EE9C724101010100530000A2FF0020DD2082014C97BA9730ED44D0D70B1FE0A4F260810200D71820D70B1FED44D0D31FD3FFD15112BAF2A122F901541044F910F2A2F80001D31F3120D74A96D307D402FB00DED1A4C8CB1FCBFFC9ED54D0E2786F"
 const _V1R3CodeHex = "B5EE9C7241010101005F0000BAFF0020DD2082014C97BA218201339CBAB19C71B0ED44D0D31FD70BFFE304E0A4F260810200D71820D70B1FED44D0D31FD3FFD15112BAF2A122F901541044F910F2A2F80001D31F3120D74A96D307D402FB00DED1A4C8CB1FCBFFC9ED54B5B86E42"
-
-var (
-	_V1R1CodeBOC []byte
-	_V1R2CodeBOC []byte
-	_V1R3CodeBOC []byte
-)
-
-func init() {
-	var err error
-
-	_V1R1CodeBOC, err = hex.DecodeString(_V1R1CodeHex)
-	if err != nil {
-		panic(err)
-	}
-	_V1R2CodeBOC, err = hex.DecodeString(_V1R2CodeHex)
-	if err != nil {
-		panic(err)
-	}
-	_V1R3CodeBOC, err = hex.DecodeString(_V1R3CodeHex)
-	if err != nil {
-		panic(err)
-	}
-}
