+++ ton/wallet/v2.go
@@ -1,25 +1,5 @@
 package wallet
 
-import "encoding/hex"
-
 // https://github.com/toncenter/tonweb/blob/master/src/contract/wallet/WalletSources.md#v2-wallet
 const _V2R1CodeHex = "B5EE9C724101010100570000AAFF0020DD2082014C97BA9730ED44D0D70B1FE0A4F2608308D71820D31FD31F01F823BBF263ED44D0D31FD3FFD15131BAF2A103F901541042F910F2A2F800029320D74A96D307D402FB00E8D1A4C8CB1FCBFFC9ED54A1370BB6"
 const _V2R2CodeHex = "B5EE9C724101010100630000C2FF0020DD2082014C97BA218201339CBAB19C71B0ED44D0D31FD70BFFE304E0A4F2608308D71820D31FD31F01F823BBF263ED44D0D31FD3FFD15131BAF2A103F901541042F910F2A2F800029320D74A96D307D402FB00E8D1A4C8CB1FCBFFC9ED54044CD7A1"
-
-var (
-	_V2R1CodeBOC []byte
-	_V2R2CodeBOC []byte
-)
-
-func init() {
-	var err error
-
-	_V2R1CodeBOC, err = hex.DecodeString(_V2R1CodeHex)
-	if err != nil {
-		panic(err)
-	}
-	_V2R2CodeBOC, err = hex.DecodeString(_V2R2CodeHex)
-	if err != nil {
-		panic(err)
-	}
-}
